出現箇所をランダムに

+++ htdocs/js/enemy1.js
@@ -8,7 +8,7 @@ Enemy1 = (function(superClass) {
   function Enemy1(x, y, speed) {
     this.image = new Image();
     this.image.src = "img/enemy1.png";
-    Enemy1.__super__.constructor.call(this, new Position(x - this.image.width.half(), y - this.image.height.half(), this.image.width, this.image.height, speed), false);
+    Enemy1.__super__.constructor.call(this, new Position(x - this.image.width.half(), y - this.image.height, this.image.width, this.image.height, speed), false);
   }
 
   Enemy1.prototype.move = function() {
